# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "description": "Returns the profile information of the authenticated user",
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                },
                "example": {
                  "id": "user123",
                  "username": "john_doe",
                  "email": "john@example.com"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "description": "Update the profile information of the authenticated user",
        "operationId": "UsersController_updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              },
              "examples": {
                "example1": {
                  "summary": "Update username",
                  "value": {
                    "username": "new_username"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponseDto"
                },
                "example": {
                  "success": true,
                  "username": "new_username"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid data provided"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user profile",
        "tags": [
          "users"
        ]
      }
    },
    "/users/search": {
      "get": {
        "description": "Search for users by username (excluding current user)",
        "operationId": "UsersController_searchUsers",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "minLength": 1,
              "example": "john",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchUserResponseDto"
                  }
                },
                "example": [
                  {
                    "id": "user456",
                    "username": "john_smith"
                  },
                  {
                    "id": "user789",
                    "username": "john_doe"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search users",
        "tags": [
          "users"
        ]
      }
    },
...
    "/chats/{chatId}": {
      "get": {
        "operationId": "ChatController_viewChat",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "The ID of the chat",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat retrieved successfully. Returns FullChatDto if user is participant, PreviewChatDto otherwise",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/FullChatDto"
                    },
                    {
                      "$ref": "#/components/schemas/PreviewChatDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - cannot view private chats of others"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "View a specific chat (full or preview)",
        "tags": [
          "chats"
        ]
      }
    },
...
    "/search": {
      "get": {
        "description": "Search for users (by username/email) and groups (by title) that the authenticated user has access to",
        "operationId": "SearchController_searchChats",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "minLength": 1,
              "example": "john",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results for users and groups",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchChatsResponseDto"
                },
                "example": {
                  "users": [
                    {
                      "id": "user123",
                      "username": "john_doe"
                    },
                    {
                      "id": "user456",
                      "username": "john_smith"
                    }
                  ],
                  "groups": [
                    {
                      "id": "chat123",
                      "title": "John's Team"
                    },
                    {
                      "id": "chat456",
                      "title": "Project John"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid search query"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search users and groups",
        "tags": [
          "search"
        ]
      }
    },
    "/search/chats/{chatId}": {
      "get": {
        "description": "Search for messages containing specific text within a chat that the user is a member of",
        "operationId": "SearchController_searchInMessages",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "Chat ID to search within",
            "schema": {
              "example": "chat123",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "minLength": 1,
              "example": "hello world",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages containing the search query",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageSearchResultDto"
                  }
                },
                "example": [
                  {
                    "id": "msg123",
                    "content": "Hello world! How are you?",
                    "createdAt": "2023-01-01T12:00:00.000Z",
                    "sender": {
                      "id": "user123",
                      "username": "john_doe"
                    }
                  },
                  {
                    "id": "msg456",
                    "content": "World news today...",
                    "createdAt": "2023-01-01T11:30:00.000Z",
                    "sender": {
                      "id": "user456",
                      "username": "jane_smith"
                    }
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "403": {
            "description": "Forbidden - User is not a member of the chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search messages in a specific chat",
        "tags": [
          "search"
        ]
      }
    }
  },
  "info": {
    "title": "Chat Backend API",
    "description": "API for chat application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "user123"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john@example.com"
          }
        },
        "required": [
          "id",
          "username",
          "email"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Update operation success status",
            "example": true
          },
          "username": {
            "type": "string",
            "description": "Updated username",
            "example": "new_username"
          }
        },
        "required": [
          "success",
          "username"
        ]
      },
      "SearchUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "user456"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "jane_smith"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
...
      "UserSearchResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "user123"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "john_doe"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "GroupSearchResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Chat/Group ID",
            "example": "chat123"
          },
          "title": {
            "type": "object",
            "description": "Group title/name",
            "example": "Team Alpha"
          }
        },
        "required": [
          "id",
          "title"
        ]
      },
      "SearchChatsResponseDto": {
        "type": "object",
        "properties": {
          "users": {
            "description": "List of users matching search",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSearchResultDto"
            }
          },
          "groups": {
            "description": "List of groups matching search",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupSearchResultDto"
            }
          }
        },
        "required": [
          "users",
          "groups"
        ]
      },
      "MessageSearchResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Message ID",
            "example": "msg123"
          },
          "content": {
            "type": "string",
            "description": "Message content",
            "example": "Hello everyone!"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Message creation timestamp",
            "example": "2023-01-01T12:00:00.000Z"
          },
          "sender": {
            "description": "Message sender information",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserSearchResultDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "content",
          "createdAt",
          "sender"
        ]
      }
    }
  }
}